+++ extensions/amazon-aws/src/ec2.tsx
@@ -6,7 +6,7 @@ import {
   OpenInBrowserAction,
   Detail,
 } from "@raycast/api";
-import { useState, useEffect } from "react";
+import { useState, useEffect, useMemo } from "react";
 import AWS from "aws-sdk";
 import setupAws from "./util/setupAws";
 
@@ -70,12 +70,19 @@ function InstanceListItem(props: { instance: AWS.EC2.Instance }) {
   const name = instance.Tags?.find((t) => t.Key === "Name")?.Value?.replace(/-/g, " ");
   const preferences: Preferences = getPreferenceValues();
 
+  const subtitle = useMemo(() => {
+    if (instance.PublicIpAddress) {
+      return `${instance.InstanceType} (🔒 ${instance.PrivateIpAddress} / 🌍 ${instance.PublicIpAddress})`;
+    }
+    return `${instance.InstanceType} (🔒 ${instance.PrivateIpAddress})`;
+  }, [instance]);
+
   return (
     <List.Item
       id={instance.InstanceId}
       key={instance.InstanceId}
       title={name || "Unknown Instance name"}
-      subtitle={instance.InstanceType + " (" + instance.PublicIpAddress + " / " + instance.PrivateIpAddress + ")"}
+      subtitle={subtitle}
       icon="list-icon.png"
       accessoryTitle={instance.LaunchTime ? instance.LaunchTime.toLocaleDateString() : undefined}
       actions={
@@ -91,8 +98,10 @@ function InstanceListItem(props: { instance: AWS.EC2.Instance }) {
               instance.InstanceId
             }
           />
-          <CopyToClipboardAction title="Copy Public IP" content={instance.PublicIpAddress || ""} />
           <CopyToClipboardAction title="Copy Private IP" content={instance.PrivateIpAddress || ""} />
+          {instance.PublicIpAddress && (
+            <CopyToClipboardAction title="Copy Public IP" content={instance.PublicIpAddress} />
+          )}
         </ActionPanel>
       }
     />