+++ extensions/ip-geolocation/src/types/ip-geolocation.ts
@@ -22,6 +22,6 @@ export interface IPGeolocationReadable {
   GeoCoordinates: string; //(lat,lon)
   Timezone: string;
   AS: string;
-  ISP: string; //ispO
+  ISP: string; //isp
   Organization: string; //org
 }