+++ extensions/open-path/src/utils/common-utils.ts
@@ -3,6 +3,9 @@ import { LocalStorage } from "@raycast/api";
 import Values = LocalStorage.Values;
 
 export interface Preference extends Values {
+  trimText: boolean;
+  isShowHud: boolean;
+  fileOperation: string;
   priorityDetection: string;
   searchEngine: string;
 }
@@ -20,8 +23,13 @@ export const checkIsFile = (path: string) => {
   }
 };
 
+export function isEmail(text: string): boolean {
+  const regex = /^[\da-zA-Z_.-]+@[\da-zA-Z_.-]+([.][a-zA-Z]+){1,2}$/;
+  return regex.test(text);
+}
+
 export function isUrl(text: string): boolean {
-  const regex = /^((http|https|ftp):\/\/)?((?:[\w-]+\.)+[a-z0-9]+)((?:\/[^/?#]*)+)?(\?[^#]+)?(#.+)?$/i;
+  const regex = /^((http|https|ftp):\/\/)?((?:[\w-]+\.)+[a-z\d]+)((?:\/[^/?#]*)+)?(\?[^#]+)?(#.+)?$/i;
   return regex.test(text) || isIP(text);
 }