+++ extensions/sourcegraph/src/components/FindNotebooksCommand.tsx
@@ -2,20 +2,18 @@ import { ActionPanel, List, Action, Icon, Detail, useNavigation } from "@raycast
 import { useState, Fragment, useMemo } from "react";
 import { DateTime } from "luxon";
 import { nanoid } from "nanoid";
-import { useQuery } from "@apollo/client";
 
 import { Sourcegraph, instanceName, LinkBuilder } from "../sourcegraph";
+import {
+  useGetNotebooksQuery,
+  NotebooksOrderBy,
+  SearchNotebookFragment as SearchNotebook,
+} from "../sourcegraph/gql/operations";
+import { bold, codeBlock, inlineCode, italic, quoteBlock } from "../markdown";
+
 import { copyShortcut } from "./shortcuts";
 import { ColorDefault, ColorEmphasis, ColorPrivate } from "./colors";
 import ExpandableErrorToast from "./ExpandableErrorToast";
-import { GET_NOTEBOOKS } from "../sourcegraph/gql/queries";
-import {
-  GetNotebooksVariables,
-  GetNotebooks,
-  SearchNotebookFields as SearchNotebook,
-  NotebooksOrderBy,
-} from "../sourcegraph/gql/schema";
-import { bold, codeBlock, inlineCode, italic, quoteBlock } from "../markdown";
 
 const link = new LinkBuilder("notebooks");
 
@@ -24,11 +22,11 @@ const link = new LinkBuilder("notebooks");
  */
 export default function FindNotebooksCommand({ src }: { src: Sourcegraph }) {
   const [searchText, setSearchText] = useState("");
-  const { loading, error, data } = useQuery<GetNotebooks, GetNotebooksVariables>(GET_NOTEBOOKS, {
+  const { loading, error, data } = useGetNotebooksQuery({
     client: src.client,
     variables: {
       query: searchText,
-      orderBy: searchText ? NotebooksOrderBy.NOTEBOOK_STAR_COUNT : NotebooksOrderBy.NOTEBOOK_UPDATED_AT,
+      orderBy: searchText ? NotebooksOrderBy.NotebookStarCount : NotebooksOrderBy.NotebookUpdatedAt,
     },
   });
   const notebooks = useMemo(() => data?.notebooks.nodes, [data]);