+++ extensions/sourcegraph/src/components/ManageBatchChangesCommand.tsx
@@ -2,31 +2,22 @@ import { ActionPanel, List, Action, Icon, useNavigation, Toast, Image, Color, sh
 import { useState, Fragment, useMemo } from "react";
 import { DateTime } from "luxon";
 import { nanoid } from "nanoid";
-import { useMutation, useQuery } from "@apollo/client";
 
 import { Sourcegraph, instanceName, LinkBuilder } from "../sourcegraph";
 import {
-  BatchChangeFields as BatchChange,
-  ChangesetFields as Changeset,
-  GetChangesets,
-  GetChangesetsVariables,
-  GetBatchChanges,
-  MergeChangeset,
-  MergeChangesetVariables,
-  ReenqueueChangeset,
-  ReenqueueChangesetVariables,
-  PublishChangesetVariables,
-  PublishChangeset,
-} from "../sourcegraph/gql/schema";
+  BatchChangeFragment as BatchChange,
+  ChangesetFragment as Changeset,
+  useGetBatchChangesQuery,
+  useGetChangesetsQuery,
+  useMergeChangesetMutation,
+  usePublishChangesetMutation,
+  useReenqueueChangesetMutation,
+} from "../sourcegraph/gql/operations";
+
 import { copyShortcut, refreshShortcut, secondaryActionShortcut, tertiaryActionShortcut } from "./shortcuts";
 import ExpandableErrorToast from "./ExpandableErrorToast";
 import { propsToKeywords } from "./keywords";
-import { GET_BATCH_CHANGES, GET_CHANGESETS } from "../sourcegraph/gql/queries";
-import {
-  MERGE_CHANGESET,
-  PUBLISH_CHANGEST as PUBLISH_CHANGESET,
-  REENQUEUE_CHANGEST as REENQUEUE_CHANGESET,
-} from "../sourcegraph/gql/mutations";
+
 import { sentenceCase } from "../text";
 
 const link = new LinkBuilder("batch-changes");
@@ -44,9 +35,7 @@ export default function ManageBatchChanges({ src }: { src: Sourcegraph }) {
    */
   const [searchText, setSearchText] = useState("");
 
-  const { loading, error, data, refetch } = useQuery<GetBatchChanges>(GET_BATCH_CHANGES, {
-    client: src.client,
-  });
+  const { loading, error, data, refetch } = useGetBatchChangesQuery({ client: src.client });
   const refresh = async () => {
     await refetch();
   };
@@ -200,8 +189,8 @@ function BatchChangeItem({
 }
 
 function BatchChangeView({ batchChange, src }: { batchChange: BatchChange; src: Sourcegraph }) {
-  const { loading, error, data, refetch } = useQuery<GetChangesets, GetChangesetsVariables>(GET_CHANGESETS, {
-    client: src.client,
+  const { loading, error, data, refetch } = useGetChangesetsQuery({
+    ...src,
     variables: {
       namespace: batchChange.namespace.id,
       name: batchChange.name,
@@ -286,20 +275,9 @@ function ChangesetItem({
     toast.hide();
   }
 
-  const [mergeChangeset, { error: mergeError }] = useMutation<MergeChangeset, MergeChangesetVariables>(
-    MERGE_CHANGESET,
-    {
-      client: src.client,
-    }
-  );
-  const [reenqueueChangeset, { error: reenqueueError }] = useMutation<ReenqueueChangeset, ReenqueueChangesetVariables>(
-    REENQUEUE_CHANGESET,
-    { client: src.client }
-  );
-  const [publishChangeset, { error: publishError }] = useMutation<PublishChangeset, PublishChangesetVariables>(
-    PUBLISH_CHANGESET,
-    { client: src.client }
-  );
+  const [mergeChangeset, { error: mergeError }] = useMergeChangesetMutation(src);
+  const [reenqueueChangeset, { error: reenqueueError }] = useReenqueueChangesetMutation(src);
+  const [publishChangeset, { error: publishError }] = usePublishChangesetMutation(src);
   const error = mergeError || publishError || reenqueueError;
   if (error) {
     ExpandableErrorToast(push, "Unexpected error", "Changeset operation failed", error.message).show();
@@ -438,7 +416,7 @@ function ChangesetItem({
 
   let title = "";
   let props = {};
-  let subtitleTooltip: string | null = null;
+  let subtitleTooltip: string | null | undefined = null;
   if (changeset.__typename === "ExternalChangeset") {
     title = `${changeset.repository.name}`;
     subtitleTooltip = changeset.title;