+++ extensions/todoist/src/components/Project.tsx
@@ -68,7 +68,11 @@ function Project({ projectId }: ProjectProps): JSX.Element {
         description="How about creating one?"
         actions={
           <ActionPanel>
-            <Action.Push title="Create Task" target={<CreateTask />} />
+            <Action.Push
+              title="Create Task"
+              target={<CreateTask fromProjectId={projectId} />}
+              shortcut={{ modifiers: ["cmd", "shift"], key: "a" }}
+            />
           </ActionPanel>
         }
       />