+++ extensions/placeholder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Placeholder Changelog
 
+## [Fix text error] - 2022-06-04
+
+- Fix text error
+
+## [Add Search Placeholder Image command] - 2022-05-23
+
+- Add Search Placeholder Image command
+- Update extension icons
+
 ## [Initial Version] - 2022-05-11
 
 - Easy to use, stylish placeholders from [Lorem Picsum](https://picsum.photos)